Extract datatype tests into datatype_test.go (#716 subtask 4)
Move 10 datatype/column-change tests from runner_test.go into datatype_test.go:
- TestChangeDatatypeNoData
- TestChangeDatatypeDataLoss
- TestChangeDatatypeLossyNoAutoInc
- TestChangeDatatypeLossless
- TestChangeDatatypeLossyFailEarly
- TestNullToNotNull
- TestTpConversion
- TestEnumReorder (unbuffered + buffered)
- TestSetReorder (unbuffered + buffered)
- TestBufferedMigrationFailsGracefullyWithMinimalRBR

Conversions:
- All table setup converted to NewTestTable (removes manual DROP/RunSQL blocks)
- All NewRunner calls converted to NewTestRunner with With* options
- Data seeding converted to SeedRows where applicable
- assert converted to require (fail-fast)
- mysql.ParseDSN eliminated from all converted tests
- Enum/set tests use tt.DB instead of separate sql.Open for DML and verification

runner_test.go: 1639 -> 1052 lines (-36%)
datatype_test.go: 379 lines (new)

Fixes #721
Part of #716 (test reorganization)
