+++ pkg/migration/runner_test.go
@@ -9,7 +9,6 @@ import (
 	"strings"
 	"sync"
 	"testing"
-	"time"
 
 	"github.com/block/spirit/pkg/dbconn"
 	"github.com/block/spirit/pkg/statement"
@@ -23,60 +22,6 @@ import (
 
 // are still supported. From https://github.com/block/spirit/issues/277
 // TestDataFromBadSqlMode tests that data previously inserted like 0000-00-00 can still be migrated.
-func TestChangeDatatypeNoData(t *testing.T) {
-	t.Parallel()
-	testutils.RunSQL(t, `DROP TABLE IF EXISTS cdatatypemytable`)
-	table := `CREATE TABLE cdatatypemytable (
-		id int(11) NOT NULL AUTO_INCREMENT,
-		name varchar(255) NOT NULL,
-		b varchar(255) NOT NULL, -- should be an int
-		PRIMARY KEY (id)
-	)`
-	testutils.RunSQL(t, table)
-	cfg, err := mysql.ParseDSN(testutils.DSN())
-	assert.NoError(t, err)
-	m, err := NewRunner(&Migration{
-		Host:     cfg.Addr,
-		Username: cfg.User,
-		Password: &cfg.Passwd,
-		Database: cfg.DBName,
-		Threads:  2,
-		Table:    "cdatatypemytable",
-		Alter:    "CHANGE b b INT", //nolint: dupword
-	})
-	assert.NoError(t, err)                // everything is specified correctly.
-	assert.NoError(t, m.Run(t.Context())) // no data so no truncation is possible.
-	assert.False(t, m.usedInstantDDL)
-	assert.NoError(t, m.Close())
-}
-
-func TestChangeDatatypeDataLoss(t *testing.T) {
-	t.Parallel()
-	testutils.RunSQL(t, `DROP TABLE IF EXISTS cdatalossmytable`)
-	table := `CREATE TABLE cdatalossmytable (
-		id int(11) NOT NULL AUTO_INCREMENT,
-		name varchar(255) NOT NULL,
-		b varchar(255) NOT NULL, -- should be an int
-		PRIMARY KEY (id)
-	)`
-	testutils.RunSQL(t, table)
-	testutils.RunSQL(t, "INSERT INTO cdatalossmytable (name, b) VALUES ('a', 'b')")
-	cfg, err := mysql.ParseDSN(testutils.DSN())
-	assert.NoError(t, err)
-	m, err := NewRunner(&Migration{
-		Host:     cfg.Addr,
-		Username: cfg.User,
-		Password: &cfg.Passwd,
-		Database: cfg.DBName,
-		Threads:  2,
-		Table:    "cdatalossmytable",
-		Alter:    "CHANGE b b INT", //nolint: dupword
-	})
-	assert.NoError(t, err)              // everything is specified correctly.
-	assert.Error(t, m.Run(t.Context())) // value 'b' can no convert cleanly to int.
-	assert.NoError(t, m.Close())
-}
-
 func TestSkipDropAfterCutoverLongTableName(t *testing.T) {
 	t.Parallel()
 
@@ -255,115 +200,6 @@ func TestBadAlter(t *testing.T) {
 // Only the key=8589934592 will fail to be converted. On my system this test
 // currently runs in 0.4 seconds which is "acceptable" for chunker performance.
 // The generated number of chunks should also be very low because of prefetching.
-func TestChangeDatatypeLossyNoAutoInc(t *testing.T) {
-	t.Parallel()
-	testutils.RunSQL(t, `DROP TABLE IF EXISTS lossychange2`)
-	table := `CREATE TABLE lossychange2 (
-					id BIGINT NOT NULL,
-					name varchar(255) NOT NULL,
-					b varchar(255) NOT NULL,
-					PRIMARY KEY (id)
-				)`
-	testutils.RunSQL(t, table)
-	testutils.RunSQL(t, "INSERT INTO lossychange2 (id, name, b) VALUES (1, 'a', REPEAT('a', 200))")          // will pass in migration
-	testutils.RunSQL(t, "INSERT INTO lossychange2 (id, name, b) VALUES (8589934592, 'a', REPEAT('a', 200))") // will fail in migration
-
-	cfg, err := mysql.ParseDSN(testutils.DSN())
-	assert.NoError(t, err)
-	m, err := NewRunner(&Migration{
-		Host:     cfg.Addr,
-		Username: cfg.User,
-		Password: &cfg.Passwd,
-		Database: cfg.DBName,
-		Threads:  2,
-		Table:    "lossychange2",
-		Alter:    "CHANGE COLUMN id id INT NOT NULL auto_increment", //nolint: dupword
-	})
-	assert.NoError(t, err)
-	err = m.Run(t.Context())
-	assert.Error(t, err)
-	assert.ErrorContains(t, err, "Out of range value") // Error 1264: Out of range value for column 'id' at row 1
-	// Check that the chunker processed fewer than 500 chunks
-	_, chunksCopied, _ := m.copier.GetChunker().Progress()
-	assert.Less(t, chunksCopied, uint64(500))
-	assert.NoError(t, m.Close())
-}
-
-// TestChangeDatatypeLossy3 has a data type change that is "lossy" but
-// given the current stored data set does not cause errors.
-func TestChangeDatatypeLossless(t *testing.T) {
-	t.Parallel()
-	testutils.RunSQL(t, `DROP TABLE IF EXISTS lossychange3`)
-	table := `CREATE TABLE lossychange3 (
-				id BIGINT NOT NULL AUTO_INCREMENT,
-				name varchar(255) NOT NULL,
-				b varchar(255) NULL,
-				PRIMARY KEY (id)
-			)`
-	testutils.RunSQL(t, table)
-	testutils.RunSQL(t, "INSERT INTO lossychange3 (name, b) VALUES ('a', REPEAT('a', 200))")
-	testutils.RunSQL(t, "INSERT INTO lossychange3 (id, name, b) VALUES (8589934592, 'a', REPEAT('a', 200))")
-
-	cfg, err := mysql.ParseDSN(testutils.DSN())
-	assert.NoError(t, err)
-	m, err := NewRunner(&Migration{
-		Host:     cfg.Addr,
-		Username: cfg.User,
-		Password: &cfg.Passwd,
-		Database: cfg.DBName,
-		Threads:  2,
-		Table:    "lossychange3",
-		Alter:    "CHANGE COLUMN b b varchar(200) NOT NULL", //nolint: dupword
-	})
-	assert.NoError(t, err)
-	err = m.Run(t.Context())
-	assert.NoError(t, err) // works because there are no violations.
-	// Check that the chunker processed fewer than 500 chunks
-	_, chunksCopied, _ := m.copier.GetChunker().Progress()
-	assert.Less(t, chunksCopied, uint64(500))
-	assert.NoError(t, m.Close())
-}
-
-// TestChangeDatatypeLossyFailEarly tests a scenario where there is an error
-// immediately so the DDL should halt.
-// So if it does try to exhaustively run the DDL it will take forever:
-// [1, 8589934592] / 1000 = 8589934.592 chunks
-
-func TestChangeDatatypeLossyFailEarly(t *testing.T) {
-	t.Parallel()
-	testutils.RunSQL(t, `DROP TABLE IF EXISTS lossychange4`)
-	table := `CREATE TABLE lossychange4 (
-				id BIGINT NOT NULL AUTO_INCREMENT,
-				name varchar(255) NOT NULL,
-				b varchar(255) NULL,
-				PRIMARY KEY (id)
-			)`
-	testutils.RunSQL(t, table)
-	testutils.RunSQL(t, "INSERT INTO lossychange4 (name) VALUES ('a')")
-	testutils.RunSQL(t, "INSERT INTO lossychange4 (name, b) VALUES ('a', REPEAT('a', 200))")
-	testutils.RunSQL(t, "INSERT INTO lossychange4 (id, name, b) VALUES (8589934592, 'a', REPEAT('a', 200))")
-	cfg, err := mysql.ParseDSN(testutils.DSN())
-	assert.NoError(t, err)
-	m, err := NewRunner(&Migration{
-		Host:     cfg.Addr,
-		Username: cfg.User,
-		Password: &cfg.Passwd,
-		Database: cfg.DBName,
-		Threads:  2,
-		Table:    "lossychange4",
-		Alter:    "CHANGE COLUMN b b varchar(255) NOT NULL", //nolint: dupword
-	})
-	assert.NoError(t, err)
-	err = m.Run(t.Context())
-	assert.Error(t, err) // there is a violation where row 1 is NULL
-	assert.NoError(t, m.Close())
-}
-
-// TestAddUniqueIndex is a really interesting test *because* resuming from checkpoint
-// will cause duplicate key errors. It's not straight-forward to differentiate between
-// duplicate errors from a resume, and a constraint violation. So what we do is:
-// 0) *FORCE* checksum to be enabled (regardless now, its always on)
-// the migration is complete, but no _chkpnt or _new or _old table.
 func TestForRemainingTableArtifacts(t *testing.T) {
 	t.Parallel()
 	testutils.RunSQL(t, `DROP TABLE IF EXISTS remainingtbl, _remainingtbl_new, _remainingtbl_old, _remainingtbl_chkpnt`)
@@ -413,85 +249,6 @@ func TestDefaultPort(t *testing.T) {
 	assert.NoError(t, err)
 	assert.Equal(t, "localhost:3306", m.migration.Host)
 	m.SetLogger(slog.Default())
-}
-
-func TestNullToNotNull(t *testing.T) {
-	t.Parallel()
-	testutils.RunSQL(t, `DROP TABLE IF EXISTS autodatetime`)
-	table := `CREATE TABLE autodatetime (
-		id INT NOT NULL AUTO_INCREMENT,
-		created_at DATETIME(3) NULL,
-		PRIMARY KEY (id)
-	)`
-	testutils.RunSQL(t, table)
-	testutils.RunSQL(t, `INSERT INTO autodatetime (created_at) VALUES (NULL)`)
-	cfg, err := mysql.ParseDSN(testutils.DSN())
-	assert.NoError(t, err)
-
-	m, err := NewRunner(&Migration{
-		Host:     cfg.Addr,
-		Username: cfg.User,
-		Password: &cfg.Passwd,
-		Database: cfg.DBName,
-		Threads:  2,
-		Table:    "autodatetime",
-		Alter:    "modify column created_at datetime(3) not null default current_timestamp(3)",
-	})
-	assert.NoError(t, err)
-	err = m.Run(t.Context())
-	assert.Error(t, err)
-	assert.ErrorContains(t, err, "Column 'created_at' cannot be null")
-	assert.NoError(t, m.Close())
-}
-
-func TestTpConversion(t *testing.T) {
-	testutils.RunSQL(t, "DROP TABLE IF EXISTS tpconvert")
-	testutils.RunSQL(t, `CREATE TABLE tpconvert (
-	id bigint NOT NULL AUTO_INCREMENT primary key,
-	created_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-	updated_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-	issued_at timestamp NULL DEFAULT NULL,
-	activated_at timestamp NULL DEFAULT NULL,
-	deactivated_at timestamp NULL DEFAULT NULL,
-	intasstring varchar(255) NULL DEFAULT NULL,
-	floatcol FLOAT NULL DEFAULT NULL
-	)`)
-	testutils.RunSQL(t, `INSERT INTO tpconvert (created_at, updated_at, issued_at, activated_at, deactivated_at, intasstring, floatcol) VALUES
-	('2023-05-18 09:28:46', '2023-05-18 09:33:27', '2023-05-18 09:28:45', '2023-05-18 09:28:45', NULL, '0001', 9.3),
-	('2023-05-18 09:34:38', '2023-05-24 07:38:25', '2023-05-18 09:34:37', '2023-05-18 09:34:37', '2023-05-24 07:38:25', '10', 9.3),
-	('2023-05-24 07:34:36', '2023-05-24 07:34:36', '2023-05-24 07:34:35', NULL, null, '01234', 9.3),
-	('2023-05-24 07:41:05', '2023-05-25 06:15:37', '2023-05-24 07:41:04', '2023-05-24 07:41:04', '2023-05-25 06:15:37', '10', 2.2),
-	('2023-05-25 06:17:30', '2023-05-25 06:17:30', '2023-05-25 06:17:29', '2023-05-25 06:17:29', NULL, '10', 9.3),
-	('2023-05-25 06:18:33', '2023-05-25 06:41:13', '2023-05-25 06:18:32', '2023-05-25 06:18:32', '2023-05-25 06:41:13', '10', 1.1),
-	('2023-05-25 06:24:23', '2023-05-25 06:24:23', '2023-05-25 06:24:22', NULL, null, '10', 9.3),
-	('2023-05-25 06:41:35', '2023-05-28 23:45:09', '2023-05-25 06:41:34', '2023-05-25 06:41:34', '2023-05-28 23:45:09', '10', 9.3),
-	('2023-05-25 06:44:41', '2023-05-28 23:45:03', '2023-05-25 06:44:40', '2023-05-25 06:46:48', '2023-05-28 23:45:03', '10', 9.3),
-	('2023-05-26 06:24:24', '2023-05-28 23:45:01', '2023-05-26 06:24:23', '2023-05-26 06:24:42', '2023-05-28 23:45:01', '10', 9.3),
-	('2023-05-28 23:46:07', '2023-05-29 00:57:55', '2023-05-28 23:46:05', '2023-05-28 23:46:05', NULL, '10', 9.3),
-	('2023-05-28 23:53:34', '2023-05-29 00:57:56', '2023-05-28 23:53:33', '2023-05-28 23:58:09', NULL, '10', 9.3);`)
-
-	cfg, err := mysql.ParseDSN(testutils.DSN())
-	assert.NoError(t, err)
-
-	m, err := NewRunner(&Migration{
-		Host:     cfg.Addr,
-		Username: cfg.User,
-		Password: &cfg.Passwd,
-		Database: cfg.DBName,
-		Threads:  2,
-		Table:    "tpconvert",
-		Alter: `MODIFY COLUMN created_at TIMESTAMP(6) NOT NULL DEFAULT CURRENT_TIMESTAMP(6),
-		MODIFY COLUMN updated_at TIMESTAMP(6) NOT NULL DEFAULT CURRENT_TIMESTAMP(6) ON UPDATE CURRENT_TIMESTAMP(6),
-		MODIFY COLUMN issued_at TIMESTAMP(6) NULL,
-		MODIFY COLUMN activated_at TIMESTAMP(6) NULL,
-		MODIFY COLUMN deactivated_at TIMESTAMP(6) NULL,
-		MODIFY COLUMN intasstring INT NULL DEFAULT NULL
-		`,
-	})
-	assert.NoError(t, err)
-	err = m.Run(t.Context())
-	assert.NoError(t, err)
-	assert.NoError(t, m.Close())
 }
 
 func TestSkipDropAfterCutover(t *testing.T) {
@@ -1247,350 +1004,6 @@ func TestDSN(t *testing.T) {
 //
 // This test exercises both the copier path (initial data) and the binlog
 // replay path (concurrent DML during migration) to verify correctness.
-func TestEnumReorder(t *testing.T) {
-	t.Run("unbuffered", func(t *testing.T) {
-		testEnumReorder(t, false)
-	})
-	t.Run("buffered", func(t *testing.T) {
-		if testutils.IsMinimalRBRTestRunner(t) {
-			t.Skip("Skipping buffered copy test because binlog_row_image is not FULL or binlog_row_value_options is not empty")
-		}
-		testEnumReorder(t, true)
-	})
-}
-
-func testEnumReorder(t *testing.T, enableBuffered bool) {
-	db, err := sql.Open("mysql", testutils.DSN())
-	require.NoError(t, err)
-	defer utils.CloseAndLog(db)
-
-	testutils.RunSQL(t, `DROP TABLE IF EXISTS enumreorder, _enumreorder_new, _enumreorder_chkpnt`)
-	testutils.RunSQL(t, `CREATE TABLE enumreorder (
-		id int NOT NULL AUTO_INCREMENT PRIMARY KEY,
-		status ENUM('active', 'inactive', 'pending') NOT NULL
-	)`)
-
-	// Insert enough initial data so the copy phase takes a measurable amount of time,
-	// giving the concurrent DML goroutine a window to inject changes that will be
-	// captured by the binlog subscription and replayed via the flush path.
-	testutils.RunSQL(t, `INSERT INTO enumreorder (status) VALUES ('active'), ('inactive'), ('pending')`)
-	testutils.RunSQL(t, `INSERT INTO enumreorder (status) SELECT status FROM enumreorder`) // 6
-	testutils.RunSQL(t, `INSERT INTO enumreorder (status) SELECT status FROM enumreorder`) // 12
-	testutils.RunSQL(t, `INSERT INTO enumreorder (status) SELECT status FROM enumreorder`) // 24
-	testutils.RunSQL(t, `INSERT INTO enumreorder (status) SELECT status FROM enumreorder`) // 48
-	testutils.RunSQL(t, `INSERT INTO enumreorder (status) SELECT status FROM enumreorder`) // 96
-	testutils.RunSQL(t, `INSERT INTO enumreorder (status) SELECT status FROM enumreorder`) // 192
-	testutils.RunSQL(t, `INSERT INTO enumreorder (status) SELECT status FROM enumreorder`) // 384
-	testutils.RunSQL(t, `INSERT INTO enumreorder (status) SELECT status FROM enumreorder`) // 768
-	testutils.RunSQL(t, `INSERT INTO enumreorder (status) SELECT status FROM enumreorder`) // 1536
-	testutils.RunSQL(t, `INSERT INTO enumreorder (status) SELECT status FROM enumreorder`) // 3072
-	testutils.RunSQL(t, `INSERT INTO enumreorder (status) SELECT status FROM enumreorder`) // 6144
-
-	cfg, err := mysql.ParseDSN(testutils.DSN())
-	require.NoError(t, err)
-
-	// Use NewRunner so we can access the runner's status to synchronize
-	// concurrent DML injection during the copy phase.
-	m, err := NewRunner(&Migration{
-		Host:            cfg.Addr,
-		Username:        cfg.User,
-		Password:        &cfg.Passwd,
-		Database:        cfg.DBName,
-		Threads:         1,
-		TargetChunkTime: 100 * time.Millisecond,
-		Table:           "enumreorder",
-		Alter:           "MODIFY COLUMN status ENUM('pending', 'active', 'inactive') NOT NULL",
-		Buffered:        enableBuffered,
-	})
-	require.NoError(t, err)
-
-	// Open a separate DB connection for DML that won't interfere with test assertions.
-	// Spirit's force-kill mechanism may kill DML connections during the checksum/lock
-	// phase, so we must not use testutils.RunSQL (which asserts no error).
-	dmlDB, err := sql.Open("mysql", testutils.DSN())
-	require.NoError(t, err)
-	defer utils.CloseAndLog(dmlDB)
-
-	// Run concurrent DML in a goroutine to exercise the binlog replay path.
-	// These inserts happen while the copier is running, so they'll be captured
-	// by the binlog subscription and flushed via deltaMap (unbuffered) or
-	// bufferedMap (buffered).
-	ctx, cancel := context.WithCancel(t.Context())
-	defer cancel()
-
-	dmlDone := make(chan struct{})
-	go func() {
-		defer close(dmlDone)
-		// Wait until we're in the copy phase
-		for m.status.Get() < status.CopyRows {
-			time.Sleep(time.Millisecond)
-			if ctx.Err() != nil {
-				return
-			}
-		}
-		// Insert rows with each ENUM value during the copy phase.
-		// These will be picked up by the binlog and replayed.
-		// We ignore errors because Spirit may kill our connection during
-		// the checksum/lock phase (force-kill of blocking transactions).
-		for i := 0; i < 50; i++ {
-			if ctx.Err() != nil {
-				return
-			}
-			_, _ = dmlDB.ExecContext(ctx, `INSERT INTO enumreorder (status) VALUES ('active')`)
-			_, _ = dmlDB.ExecContext(ctx, `INSERT INTO enumreorder (status) VALUES ('inactive')`)
-			_, _ = dmlDB.ExecContext(ctx, `INSERT INTO enumreorder (status) VALUES ('pending')`)
-			// Also update some existing rows to exercise the update path
-			_, _ = dmlDB.ExecContext(ctx, fmt.Sprintf(`UPDATE enumreorder SET status = 'active' WHERE id = %d`, i*3+1))
-			_, _ = dmlDB.ExecContext(ctx, fmt.Sprintf(`UPDATE enumreorder SET status = 'inactive' WHERE id = %d`, i*3+2))
-			_, _ = dmlDB.ExecContext(ctx, fmt.Sprintf(`UPDATE enumreorder SET status = 'pending' WHERE id = %d`, i*3+3))
-		}
-	}()
-
-	// Run the migration. For unbuffered mode this should succeed with correct data.
-	// For buffered mode, the preflight check should refuse the ENUM reorder
-	// because the binlog replay path uses integer ordinals that would cause corruption.
-	migrationErr := m.Run(ctx)
-	cancel()
-	<-dmlDone
-	assert.NoError(t, m.Close())
-
-	if enableBuffered {
-		// Buffered mode: the preflight check should refuse this migration because
-		// ENUM reordering is unsafe when the binlog replay path uses integer ordinals.
-		// Accepting this preflight error as a "pass" — it's not corruption, it's
-		// Spirit correctly refusing a change it can't handle safely.
-		require.Error(t, migrationErr)
-		assert.ErrorContains(t, migrationErr, "unsafe ENUM value reorder")
-		return
-	}
-
-	// Unbuffered mode: migration should succeed and data should be correct.
-	require.NoError(t, migrationErr)
-
-	// Verify that every row has a valid ENUM string value and that
-	// no ordinal-based corruption occurred.
-	var activeCount, inactiveCount, pendingCount int
-	err = db.QueryRowContext(t.Context(), `SELECT
-			SUM(status = 'active'),
-			SUM(status = 'inactive'),
-			SUM(status = 'pending')
-			FROM enumreorder`).Scan(&activeCount, &inactiveCount, &pendingCount)
-	require.NoError(t, err)
-
-	var totalCount int
-	err = db.QueryRowContext(t.Context(), `SELECT COUNT(*) FROM enumreorder`).Scan(&totalCount)
-	require.NoError(t, err)
-
-	// Every row must map to one of the three valid ENUM string values.
-	assert.Equal(t, totalCount, activeCount+inactiveCount+pendingCount,
-		"all rows should have a valid ENUM value (no empty strings from ordinal corruption)")
-	assert.Greater(t, activeCount, 0, "should have 'active' rows")
-	assert.Greater(t, inactiveCount, 0, "should have 'inactive' rows")
-	assert.Greater(t, pendingCount, 0, "should have 'pending' rows")
-}
-
-// TestSetReorder mirrors TestEnumReorder but for SET columns.
-// SET values are stored as bitmasks in the binlog, so reordering has the same
-// corruption risk as ENUM in buffered mode.
-func TestSetReorder(t *testing.T) {
-	t.Run("unbuffered", func(t *testing.T) {
-		testSetReorder(t, false)
-	})
-	t.Run("buffered", func(t *testing.T) {
-		if testutils.IsMinimalRBRTestRunner(t) {
-			t.Skip("Skipping buffered copy test because binlog_row_image is not FULL or binlog_row_value_options is not empty")
-		}
-		testSetReorder(t, true)
-	})
-}
-
-func testSetReorder(t *testing.T, enableBuffered bool) {
-	db, err := sql.Open("mysql", testutils.DSN())
-	require.NoError(t, err)
-	defer utils.CloseAndLog(db)
-
-	testutils.RunSQL(t, `DROP TABLE IF EXISTS setreorder, _setreorder_new, _setreorder_chkpnt`)
-	testutils.RunSQL(t, `CREATE TABLE setreorder (
-		id int NOT NULL AUTO_INCREMENT PRIMARY KEY,
-		perms SET('read', 'write', 'execute') NOT NULL
-	)`)
-
-	testutils.RunSQL(t, `INSERT INTO setreorder (perms) VALUES ('read'), ('write'), ('execute'), ('read,write'), ('read,execute'), ('write,execute'), ('read,write,execute')`)
-	testutils.RunSQL(t, `INSERT INTO setreorder (perms) SELECT perms FROM setreorder`) // 14
-	testutils.RunSQL(t, `INSERT INTO setreorder (perms) SELECT perms FROM setreorder`) // 28
-	testutils.RunSQL(t, `INSERT INTO setreorder (perms) SELECT perms FROM setreorder`) // 56
-	testutils.RunSQL(t, `INSERT INTO setreorder (perms) SELECT perms FROM setreorder`) // 112
-	testutils.RunSQL(t, `INSERT INTO setreorder (perms) SELECT perms FROM setreorder`) // 224
-	testutils.RunSQL(t, `INSERT INTO setreorder (perms) SELECT perms FROM setreorder`) // 448
-	testutils.RunSQL(t, `INSERT INTO setreorder (perms) SELECT perms FROM setreorder`) // 896
-	testutils.RunSQL(t, `INSERT INTO setreorder (perms) SELECT perms FROM setreorder`) // 1792
-	testutils.RunSQL(t, `INSERT INTO setreorder (perms) SELECT perms FROM setreorder`) // 3584
-	testutils.RunSQL(t, `INSERT INTO setreorder (perms) SELECT perms FROM setreorder`) // 7168
-
-	cfg, err := mysql.ParseDSN(testutils.DSN())
-	require.NoError(t, err)
-
-	m, err := NewRunner(&Migration{
-		Host:            cfg.Addr,
-		Username:        cfg.User,
-		Password:        &cfg.Passwd,
-		Database:        cfg.DBName,
-		Threads:         1,
-		TargetChunkTime: 100 * time.Millisecond,
-		Table:           "setreorder",
-		Alter:           "MODIFY COLUMN perms SET('execute', 'read', 'write') NOT NULL",
-		Buffered:        enableBuffered,
-	})
-	require.NoError(t, err)
-
-	dmlDB, err := sql.Open("mysql", testutils.DSN())
-	require.NoError(t, err)
-	defer utils.CloseAndLog(dmlDB)
-
-	ctx, cancel := context.WithCancel(t.Context())
-	defer cancel()
-
-	dmlDone := make(chan struct{})
-	go func() {
-		defer close(dmlDone)
-		for m.status.Get() < status.CopyRows {
-			time.Sleep(time.Millisecond)
-			if ctx.Err() != nil {
-				return
-			}
-		}
-		for i := 0; i < 50; i++ {
-			if ctx.Err() != nil {
-				return
-			}
-			_, _ = dmlDB.ExecContext(ctx, `INSERT INTO setreorder (perms) VALUES ('read')`)
-			_, _ = dmlDB.ExecContext(ctx, `INSERT INTO setreorder (perms) VALUES ('write,execute')`)
-			_, _ = dmlDB.ExecContext(ctx, `INSERT INTO setreorder (perms) VALUES ('read,write,execute')`)
-			_, _ = dmlDB.ExecContext(ctx, fmt.Sprintf(`UPDATE setreorder SET perms = 'read,write' WHERE id = %d`, i*3+1))
-			_, _ = dmlDB.ExecContext(ctx, fmt.Sprintf(`UPDATE setreorder SET perms = 'execute' WHERE id = %d`, i*3+2))
-			_, _ = dmlDB.ExecContext(ctx, fmt.Sprintf(`UPDATE setreorder SET perms = 'read,write,execute' WHERE id = %d`, i*3+3))
-		}
-	}()
-
-	migrationErr := m.Run(ctx)
-	cancel()
-	<-dmlDone
-	assert.NoError(t, m.Close())
-
-	// Both buffered and unbuffered modes should refuse SET reordering.
-	// In buffered mode, the binlog replay path uses bitmask ordinals that corrupt data.
-	// In unbuffered mode, the data is actually correct but MySQL outputs SET members
-	// in definition order, so the string representation changes (e.g. "read,execute"
-	// becomes "execute,read"), causing Spirit's CRC32 checksum to always fail.
-	require.Error(t, migrationErr)
-	assert.ErrorContains(t, migrationErr, "unsafe SET value reorder")
-}
-
-// TestBufferedMigrationFailsGracefullyWithMinimalRBR verifies that a buffered
-// migration fails gracefully when it receives minimal RBR events. The buffered
-// mode uses a buffered applier which requires full row images. We simulate a
-// rogue session that has SET binlog_row_image = 'minimal' at the session level,
-// which causes its DML to produce minimal row images in the binlog even though
-// the global setting is FULL.
-func TestBufferedMigrationFailsGracefullyWithMinimalRBR(t *testing.T) {
-	if testutils.IsMinimalRBRTestRunner(t) {
-		t.Skip("Skipping test for minimal RBR test runner (global setting already minimal)")
-	}
-
-	db, err := sql.Open("mysql", testutils.DSN())
-	require.NoError(t, err)
-	defer utils.CloseAndLog(db)
-
-	testutils.RunSQL(t, `DROP TABLE IF EXISTS minrbr_buffered, _minrbr_buffered_new, _minrbr_buffered_chkpnt`)
-	testutils.RunSQL(t, `CREATE TABLE minrbr_buffered (
-		id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
-		name VARCHAR(255) NOT NULL,
-		val INT NOT NULL DEFAULT 0
-	)`)
-
-	// Insert some data. The test throttler will slow the copier down,
-	// giving the repl client time to process the minimal-RBR events.
-	testutils.RunSQL(t, `INSERT INTO minrbr_buffered (name, val) VALUES ('seed', 1)`)
-	for range 8 {
-		testutils.RunSQL(t, `INSERT INTO minrbr_buffered (name, val) SELECT CONCAT(name, '-', id), val FROM minrbr_buffered`)
-	}
-
-	cfg, err := mysql.ParseDSN(testutils.DSN())
-	require.NoError(t, err)
-
-	// Open a dedicated connection with session-level minimal RBR.
-	// DML on this connection will produce minimal row images in the binlog.
-	minimalDB, err := sql.Open("mysql", testutils.DSN())
-	require.NoError(t, err)
-	defer utils.CloseAndLog(minimalDB)
-
-	// Force a single connection so the session variable sticks.
-	minimalDB.SetMaxOpenConns(1)
-	_, err = minimalDB.ExecContext(t.Context(), "SET binlog_row_image = 'MINIMAL'")
-	require.NoError(t, err)
-
-	ctx, cancel := context.WithCancel(t.Context())
-	defer cancel()
-
-	m, err := NewRunner(&Migration{
-		Host:             cfg.Addr,
-		Username:         cfg.User,
-		Password:         &cfg.Passwd,
-		Database:         cfg.DBName,
-		Threads:          1,
-		TargetChunkTime:  100 * time.Millisecond,
-		Table:            "minrbr_buffered",
-		Alter:            "ENGINE=InnoDB",
-		Buffered:         true,
-		useTestThrottler: true, // slows the copier so the repl client has time to see minimal events
-	})
-	require.NoError(t, err)
-
-	// Run the migration in a goroutine so we can inject minimal-RBR writes
-	// once the copy phase has started.
-	var migrationErr error
-	migrationDone := make(chan struct{})
-	go func() {
-		defer close(migrationDone)
-		migrationErr = m.Run(ctx)
-	}()
-
-	// Wait until the migration is in the copy phase, then start writing
-	// with minimal RBR. This ensures the repl client is actively reading
-	// binlog events when the minimal row images arrive.
-	waitForStatus(t, m, status.CopyRows)
-
-	// Continuously write using the minimal-RBR session during the copy phase.
-	// The test throttler slows the copier, so these writes will be picked up
-	// by the binlog subscription before the copy phase finishes.
-	var writerWg sync.WaitGroup
-	writerWg.Add(1)
-	go func() {
-		defer writerWg.Done()
-		for i := 0; ; i++ {
-			select {
-			case <-ctx.Done():
-				return
-			default:
-				_, _ = minimalDB.ExecContext(ctx, `UPDATE minrbr_buffered SET val = val + 1 WHERE id = ?`, (i%100)+1)
-			}
-		}
-	}()
-
-	// Wait for the migration to complete (it should fail), then stop the writer.
-	<-migrationDone
-	cancel()
-	writerWg.Wait()
-	assert.NoError(t, m.Close())
-
-	// The migration should fail because the runtime check detects minimal RBR
-	// events while a buffered applier is in use. The repl client cancels
-	// the caller's context, so the error may be context.Canceled or may
-	// contain the original "minimal RBR" message depending on which
-	// operation observes the cancellation first.
-	require.Error(t, migrationErr)
-}
-
 func TestMigrationWithSQLCommentsInStatement(t *testing.T) {
 	// This test verifies that Spirit correctly handles SQL comments
 	// prepended to ALTER TABLE statements when using the Statement field
